Manage favorite meals with context api
