+++ meals-app/components/MealsList/MealItem.js
@@ -8,7 +8,7 @@ import {
   Platform,
 } from "react-native";
 import { useNavigation } from "@react-navigation/native";
-import MealDetail from "./MealDetails";
+import MealDetail from "../MealDetails";
 
 function MealItem({
   id,
