+++ meals-app/screens/FavoritesScreen.js
@@ -1,14 +1,37 @@
-import React from 'react'
-import { View } from 'react-native'
+import React, { useContext } from 'react'
+import { View, Text, StyleSheet } from 'react-native';
+
+
+import MealsList from '../components/MealsList/MealsList'
+import { FavoriteContext } from '../store/context/Favarites-context'
+import { MEALS } from '../data/dummy-data';
 
 function FavoritesScreen() {
-  return (
-    <View>
-        <Text>
-            My Favorite
-        </Text>
+  const favoriteMealsContext = useContext(FavoriteContext);
+
+  const favoriteMeal = MEALS.filter(meal =>favoriteMealsContext.ids.includes(meal.id))
+
+  if(favoriteMeal.length === 0){
+    return <View style={styles.rootContainer}>
+      <Text style={styles.text}>You have no favorite meals yet</Text>
     </View>
+  }
+  return (
+    <MealsList items={favoriteMeal}/>
   )
 }
 
-export default FavoritesScreen
+export default FavoritesScreen;
+
+const styles = StyleSheet.create({
+  rootContainer: {
+    flex: 1,
+    justifyContent: "center",
+    alignItems: "center",
+  },
+  text: {
+    fontSize: 18,
+    fontWeight: "bold",
+    color: "white"
+  }
+})
